# server/src/config/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Movio AI API v4.1",
    "version": "4.1.0",
    "description": "电商AI SaaS系统接口文档 — 图片/视频/批量/商业化/管理后台 — v4.1 重构版"
  },
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "开发环境"
    },
    {
      "url": "https://api.movio.ai",
      "description": "生产环境"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/api/points/transactions": {
      "get": {
        "tags": [
          "Points v4"
        ],
        "summary": "积分流水",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "分页流水列表"
          }
        }
      }
    },
    "/api/points/redeem": {
      "post": {
        "tags": [
          "Points v4"
        ],
        "summary": "积分兑换点数",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "points"
                ],
                "properties": {
                  "points": {
                    "type": "integer",
                    "minimum": 1
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "兑换成功，返回redeemed_credits"
          }
        }
      }
    },
    "/api/points/earn": {
      "post": {
        "tags": [
          "Points v4"
        ],
        "summary": "管理后台手动发放积分",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "user_id",
                  "amount"
                ],
                "properties": {
                  "user_id": {
                    "type": "integer"
                  },
                  "amount": {
                    "type": "integer"
                  },
                  "remark": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "积分发放成功"
          }
        }
      }
    },
...
  }
}